requirements.txt
fastapi
uvicorn[standard]

pydantic
python-jose
passlib[bcrypt]

sqlalchemy
asyncpg
redis

sentence-transformers
faiss-cpu

python-multipart
requests
python-dotenv
gunicorn
PyPDF2